Rename hyperopt interface and resolver
